observes\feat(job): #3038.8 format dates and add index
Former-commit-id: 99ae66df56787282bffe0ece3d143fdd1ad89ba5

+++ observes/singer/streamer_dynamodb/streamer_dynamodb/__init__.py
@@ -112,7 +112,7 @@ def deserialize(object_: Any) -> Any:
     :rtype: Any.
     """
     if isinstance(object_, Decimal):
-        object_ = int(object_)
+        object_ = float(object_)
     elif isinstance(object_, dict):
         for key, value in object_.items():
             object_[key] = deserialize(value)
